DivisionGrid; buttons for divisions

+++ src/flagmaker/Divisions/Division.java
@@ -8,9 +8,9 @@ import javafx.scene.paint.Color;
 public abstract class Division
 {
 	public Color[] Colors;
-	public Double[] Values;
+	public int[] Values;
 	
-	protected Division(Color[] colors, Double[] values)
+	protected Division(Color[] colors, int[] values)
 	{
 		Colors = colors;
 		Values = values;
@@ -19,6 +19,6 @@ public abstract class Division
 	public abstract String Name();
 	public abstract void Draw(Pane canvas);
 	public abstract void SetColors(Color[] colors);
-	public abstract void SetValues(Double[] values);
+	public abstract void SetValues(int[] values);
 	public abstract String ExportSvg(int width, int height);
 }
